library reorganization
