+++ libs/confetti/types.ts
@@ -1,0 +1,17 @@
+import { Coord } from '../utilities/CanvasUtils';   
+
+export enum EAnimationState {
+    READY,
+    LOADED,
+    STARTED,
+    ENDED,
+}
+
+export type ConfettiParticleConfig = {
+    context: CanvasRenderingContext2D;
+    p0: Coord;
+    p1: Coord;
+    p2: Coord;
+    p3: Coord;
+    baseColor?: string;
+}
